Factor out OrderingPositionEvaluator#getBucketScore
This was originally part of a feature to cache bucket scores. But the
bit twiddling logic is apparenlty fast enough that the cache overhead
slows it down by half. So I took out the cache. I even tried with a
straight HashMap, and it slowed down the app by 20%. I wonder if it
would be a net win if we used deeper buckets? But we can't handle
deeper buckets yet, so we won't look that way yet.