Fix ordering bug
Turns out PriorityQueues in Java are intrinsically min-heaps, and I was
treating it with max-heap semantics. So when it took 1700 moves to find
the right answer, it's kind of amazing we found it at all.

With this new change, we're processing slightly fewer moves per second
(~800K vs 1mm prior), but it solves in 80 moves, and I think the model
will handle larger depths more robustly, too.